[BE] MOVE: 파일 구조 정리 #174
