+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/config/JacksonConfig.java
@@ -1,4 +1,4 @@
-package kickzo.stomp_chat.utils;
+package kickzo.stomp_chat.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
