+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/dto/room/ChatMessage.java
@@ -1,3 +1,3 @@
-package kickzo.stomp_chat.dto;
+package kickzo.stomp_chat.dto.room;
 
 public record ChatMessage(long roomId, long userId, String content, String message) {}
