+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/dto/room/RoleChangeEvent.java
@@ -1,4 +1,4 @@
-package kickzo.stomp_chat.dto;
+package kickzo.stomp_chat.dto.room;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class PlaylistItem {
-	private String url;
-	private int order;
+public class RoleChangeEvent {
+	private Long roomId;
+	private Long targetUserId;
+	private int newRole;
 }
