+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/dto/user/FriendNotification.java
@@ -1,4 +1,4 @@
-package kickzo.stomp_chat.dto;
+package kickzo.stomp_chat.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
