+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/service/MessagingService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import kickzo.stomp_chat.dto.ChatMessage;
+import kickzo.stomp_chat.dto.room.ChatMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
