+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/service/RoomEventHandler.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import kickzo.stomp_chat.dto.NewUserJoinEvent;
-import kickzo.stomp_chat.dto.RoleChangeEvent;
-import kickzo.stomp_chat.dto.RoomData;
+import kickzo.stomp_chat.dto.room.NewUserJoinEvent;
+import kickzo.stomp_chat.dto.room.RoleChangeEvent;
+import kickzo.stomp_chat.dto.room.RoomData;
 import kickzo.stomp_chat.dto.RoomEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
